Fix: Move client server into public, update neccessary files for PR to be complete
