+++ .env.example
@@ -6,3 +6,9 @@ GITHUB_PERSONAL_ACCESS_TOKEN=
 
 # Supply base URL
 GITHUB_API_URL=
+
+# For client only: Supply full backend url
+BACKEND_URL=
+
+# For client only: Supply port for client's server
+CLIENT_PORT=
