+++ .gitignore
@@ -1,3 +1,4 @@
 .vscode
 node_modules
 .env
+config.generated.json
