+++ package.json
@@ -8,7 +8,7 @@
   },
   "scripts": {
     "start": "nodemon server.js",
-    "start:client": "node public/lib/generate.config.js && node client-server.js"
+    "start:client": "node public/lib/generate.config.js && node public/server.js"
   },
   "devDependencies": {
     "nodemon": "^2.0.22"
